pyproject.toml
[tool.poetry]
name = "itstart-backend"
version = "0.1.0"
description = "Monorepo: FastAPI core API + aiogram Telegram bot + shared domain for IT-Start project"
authors = ["IT-Start Team <dev@example.com>"]
readme = "README.md"
packages = [
    { include = "itstart_domain", from = "src" },
    { include = "itstart_core_api", from = "src" },
    { include = "itstart_tg_bot", from = "src" },
    { include = "itstart_common", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
sqlalchemy = { extras = ["asyncio"], version = "^2.0.30" }
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic = "^2.8.2"
pydantic-settings = "^2.2.1"
passlib = { extras = ["argon2"], version = "^1.7.4" }
python-gnupg = "^0.5.2"
redis = "^5.0.4"
aiogram = "^3.6.0"
celery = "^5.4.0"
kombu = "^5.4.2"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
prometheus-client = "^0.20.0"
sentry-sdk = "^2.8.0"
httpx = "^0.27.0"
tenacity = "^8.3.0"
pyotp = "^2.9.0"
openpyxl = "^3.1.5"
pgpy = "^0.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.7"
pytest-cov = "^5.0.0"
black = "^24.8.0"
isort = "^5.13.2"
ruff = "^0.5.6"
mypy = "^1.11.1"
types-redis = "^4.6.0.20240425"
pre-commit = "^3.8.0"

[tool.poetry.scripts]
itstart-core-api = "itstart_core_api.main:run"
itstart-tg-bot = "itstart_tg_bot.main:run"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py310"
[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501", "B008", "UP045", "B904"]

[tool.mypy]
python_version = "3.10"
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
strict_optional = true
exclude = ["tests"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term --cov-report=xml"

[tool.coverage.run]
omit = [
  "src/itstart_core_api/main.py",
  "src/itstart_core_api/celery_app.py",
  "src/itstart_core_api/export.py",
  "src/itstart_core_api/tag_seed.py",
  "src/itstart_tg_bot/bot.py",
  "src/itstart_tg_bot/main.py",
  "src/itstart_tg_bot/db.py",
]
